requirements.txt
Django==1.2.4
South==0.7.3
django-debug-toolbar==0.8.4
django-extensions==0.6
django-registration==0.7
ipython==0.10.1
wsgiref==0.1.2
